fix: resolve npm install dependency conflict and path issue
- Remove react-swipeable-views (^0.14.0) from dependencies
  - Package only supports React 15/16/17, incompatible with React 18.3.1
  - Causes ERESOLVE error on plain 'npm install'
  - Package is not imported anywhere in source code
  - swiper (already a dep) serves as the modern equivalent
- Add .npmrc with legacy-peer-deps=true as a safety net for
  other transitive peer dependency mismatches
- Fix SignupSuccess component path: removed trailing spaces from
  directory and file name that caused resolution issues on
  case-sensitive file systems and CI environments
- Update package-lock.json to reflect removed dependency
